Change it from com.whylogs.core to whylogs.core Means moving files, changing import field
