--- java/core/src/main/java/com/whylogs/core/FrequentItemsErrorType.java
@@ -1,6 +1,0 @@
-package com.whylogs.core;
-
-public enum FrequentItemsErrorType {
-  NO_FALSE_POSITIVES,
-  NO_FALSE_NEGATIVES
-}
