--- java/core/src/main/java/com/whylogs/core/SummaryConfig.java
@@ -1,14 +1,0 @@
-package com.whylogs.core;
-
-import java.util.ArrayList;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@NoArgsConstructor
-public class SummaryConfig {
-  private ArrayList<String> disabledMetrics;
-  private FrequentItemsErrorType frequentItemsErrorType = FrequentItemsErrorType.NO_FALSE_POSITIVES;
-  private int frequentItemsLimit = -1;
-  private int hll_stddev = 1;
-}
