--- java/core/src/main/java/com/whylogs/core/errors/DeserializationError.java
@@ -1,8 +1,0 @@
-package com.whylogs.core.errors;
-
-/** Exception raised when deserializing data. */
-public class DeserializationError extends Error {
-  public DeserializationError(String message) {
-    super(message);
-  }
-}
