--- java/core/src/main/java/com/whylogs/core/errors/Error.java
@@ -1,8 +1,0 @@
-package com.whylogs.core.errors;
-
-/** Base error type for this module. */
-public class Error extends Exception {
-  public Error(String message) {
-    super(message);
-  }
-}
