--- java/core/src/main/java/com/whylogs/core/errors/LoggingError.java
@@ -1,8 +1,0 @@
-package com.whylogs.core.errors;
-
-/** Exception when an error happens during logging. */
-public class LoggingError extends Error {
-  public LoggingError(String message) {
-    super(message);
-  }
-}
