--- java/core/src/main/java/com/whylogs/core/errors/SerializationError.java
@@ -1,8 +1,0 @@
-package com.whylogs.core.errors;
-
-/** Exception raised when serializing data. */
-public class SerializationError extends Error {
-  public SerializationError(String message) {
-    super(message);
-  }
-}
