--- java/core/src/main/java/com/whylogs/core/errors/UnsupportedError.java
@@ -1,8 +1,0 @@
-package com.whylogs.core.errors;
-
-/** * Exception raised when an operation is not supported. */
-public class UnsupportedError extends Error {
-  public UnsupportedError(String message) {
-    super(message);
-  }
-}
