--- java/core/src/main/java/com/whylogs/core/metrics/components/IntegralComponent.java
@@ -1,16 +1,0 @@
-package com.whylogs.core.metrics.components;
-
-import lombok.EqualsAndHashCode;
-
-@EqualsAndHashCode(callSuper = false)
-public class IntegralComponent extends MetricComponent<Integer> {
-  // everything is the same as MetricComponent, mtype=int
-
-  public IntegralComponent() {
-    super(0);
-  }
-
-  public IntegralComponent(Integer value) {
-    super(value);
-  }
-}
