+++ java/core/src/main/java/whylogs/core/FrequentItemsErrorType.java
@@ -1,0 +1,6 @@
+package whylogs.core;
+
+public enum FrequentItemsErrorType {
+  NO_FALSE_POSITIVES,
+  NO_FALSE_NEGATIVES
+}
