+++ java/core/src/main/java/whylogs/core/errors/BadConfigError.java
@@ -1,0 +1,8 @@
+package whylogs.core.errors;
+
+/** Exception when an error is due to bad configuration. */
+public class BadConfigError extends Error {
+  public BadConfigError(String message) {
+    super(message);
+  }
+}
