+++ java/core/src/main/java/whylogs/core/errors/DeserializationError.java
@@ -1,0 +1,8 @@
+package whylogs.core.errors;
+
+/** Exception raised when deserializing data. */
+public class DeserializationError extends Error {
+  public DeserializationError(String message) {
+    super(message);
+  }
+}
