+++ java/core/src/main/java/whylogs/core/errors/Error.java
@@ -1,0 +1,8 @@
+package whylogs.core.errors;
+
+/** Base error type for this module. */
+public class Error extends Exception {
+  public Error(String message) {
+    super(message);
+  }
+}
