+++ java/core/src/main/java/whylogs/core/errors/LoggingError.java
@@ -1,0 +1,8 @@
+package whylogs.core.errors;
+
+/** Exception when an error happens during logging. */
+public class LoggingError extends Error {
+  public LoggingError(String message) {
+    super(message);
+  }
+}
