+++ java/core/src/main/java/whylogs/core/errors/SerializationError.java
@@ -1,0 +1,8 @@
+package whylogs.core.errors;
+
+/** Exception raised when serializing data. */
+public class SerializationError extends Error {
+  public SerializationError(String message) {
+    super(message);
+  }
+}
