+++ java/core/src/main/java/whylogs/core/errors/UnsupportedError.java
@@ -1,0 +1,8 @@
+package whylogs.core.errors;
+
+/** * Exception raised when an operation is not supported. */
+public class UnsupportedError extends Error {
+  public UnsupportedError(String message) {
+    super(message);
+  }
+}
