--- java/core/src/test/java/com/whylogs/core/metrics/components/TestIntegralComponent.java
@@ -1,19 +1,0 @@
-package com.whylogs.core.metrics.components;
-
-import org.testng.Assert;
-import org.testng.annotations.Test;
-
-@Test
-public class TestIntegralComponent {
-
-  @Test
-  public void test_integral() {
-    IntegralComponent component = new IntegralComponent(1);
-    Assert.assertEquals((int) component.getValue(), 1);
-
-    component = new IntegralComponent();
-    Assert.assertEquals((int) component.getValue(), 0);
-
-    Assert.assertEquals(component.getTypeId(), 0);
-  }
-}
